extract Navbar into single component
